# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Asocanitas: API",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module.<br>Karen Pabón"
  },
  "host": "http://asocanitas.herokuapp.com/",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "apiKeyAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "X-API-KEY",
      "description": "any description..."
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/user/": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "Endpoint to sign in a specific user...",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      },
      "get": {
        "tags": [],
        "description": "",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/user/{user}": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "user",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/route/": {
      "post": {
        "tags": [],
        "description": "",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "get": {
        "tags": [],
        "description": "",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/route/{user}": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "user",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  },
...
}